criacao de metodos genericos

+++ PesquisaEleitoral/PesquisaEleitoral/DTOs/Estatisticas/PerfilEleitoresDTO.cs
@@ -8,10 +8,10 @@ namespace PesquisaEleitoral.DTOs.Estatisticas
         public string Nome { get; set; } = string.Empty;
 
         public int TotalVotos { get; set; }
-        public double PorcentagemVotos { get; set; }
+        public decimal PorcentagemVotos { get; set; }
         public Dictionary<FaixaEtaria, decimal> DistribuicaoFaixaEtaria { get; set; } = null!;
         public Dictionary<ClasseSocial, decimal> DistribuicaoRenda { get; set; } = null!;
-        public Dictionary<Escolaridade, double> DistribuicaoEscolaridade { get; set; } = null!;
-        public Dictionary<Sexo, double> DistribuicaoSexo { get; set; } = null!;
+        public Dictionary<Escolaridade, decimal> DistribuicaoEscolaridade { get; set; } = null!;
+        public Dictionary<Sexo, decimal> DistribuicaoSexo { get; set; } = null!;
     }
 }
